moved controller and removed service
