--- service/grails-app/controllers/org/olf/DashboardDefinitionsController.groovy
@@ -1,14 +1,0 @@
-package org.olf
-
-import grails.gorm.multitenancy.CurrentTenant
-import groovy.util.logging.Slf4j
-
-@Slf4j
-@CurrentTenant
-class DashboardDefinitionsController  {
-  def dashboardDefinitionsService
-  public getDefinitions() {
-    respond dashboardDefinitionsService.getWidgetDefinitions();
-  }
-}
-
